+++ src/addons/display_commander/utils/general_utils.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #define ImTextureID ImU64
-#define DEBUG_LEVEL_0
 #define WIN32_LEAN_AND_MEAN
 
 #include <windows.h>